:art: ActivityMain.kt 코드정리
