+++ app/src/main/java/com/chobo/onrest/MainActivity.kt
@@ -33,244 +33,12 @@ class MainActivity : AppCompatActivity() {
     }
 }
 
-class Calender : AppCompatActivity() {
-    private lateinit var binding: CalendarBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = CalendarBinding.inflate(layoutInflater)
-        val view = binding.root
-        setContentView(view)
 
-        binding.speechbubble.setOnClickListener() {
-            startActivity(Intent(this, DiaryWrite::class.java))
-            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
-            finish()
-        }
-        binding.stripe.setOnClickListener() {
-            startActivity(Intent(this, DiaryWrite::class.java))
-            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
-            finish()
-        }
-    }
-}
-class DiaryWrite : AppCompatActivity() {
-    private lateinit var binding: DiaryWriteBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = DiaryWriteBinding.inflate(layoutInflater)
-        val view = binding.root
-        setContentView(view)
 
-        binding.write.setOnClickListener() {
-            val intent = Intent(this, YourEmotion::class.java)
-            startActivity(intent)
-            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
-            finish()
-        }
-    }
-}
-class YourEmotion : AppCompatActivity() {
-    private lateinit var binding: YourEmotionBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = YourEmotionBinding.inflate(layoutInflater)
-        val view = binding.root
-        setContentView(view)
 
-        binding.choice1.setOnClickListener() {
-            startActivity(Intent(this, QuestList::class.java))
-            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
-            finish()
-        }
-        binding.choice2.setOnClickListener() {
-            startActivity(Intent(this, EmotionChoice1::class.java))
-            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
-            finish()
-        }
-    }
-}
-class EmotionChoice1 : AppCompatActivity() {
-    private lateinit var binding: EmotionChoice1Binding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = EmotionChoice1Binding.inflate(layoutInflater)
-        val view = binding.root
-        setContentView(view)
 
-        binding.happyface.setOnClickListener() {
-            startActivity(Intent(this, EmotionChoice2Happy::class.java))
-            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
-            finish()
-        }
-        binding.angryface.setOnClickListener() {
-            startActivity(Intent(this, EmotionChoice2Angry::class.java))
-            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
-            finish()
-        }
-        binding.sadface.setOnClickListener() {
-            startActivity(Intent(this, EmotionChoice2Sad::class.java))
-            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
-            finish()
-        }
-    }
-}
-class EmotionChoice2Happy : AppCompatActivity() {
-    private lateinit var binding: EmotionChoice2HappyBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = EmotionChoice2HappyBinding.inflate(layoutInflater)
-        val view = binding.root
-        setContentView(view)
 
-        binding.choice1.setOnClickListener() {
-            startActivity(Intent(this, QuestChoice::class.java))
-            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
-            finish()
-        }
-        binding.choice2.setOnClickListener() {
-            startActivity(Intent(this, QuestChoice::class.java))
-            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
-            finish()
-        }
-        binding.choice3.setOnClickListener() {
-            startActivity(Intent(this, QuestChoice::class.java))
-            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
-            finish()
-        }
-    }
-}
-class EmotionChoice2Angry : AppCompatActivity() {
-    private lateinit var binding: EmotionChoice2AngryBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = EmotionChoice2AngryBinding.inflate(layoutInflater)
-        val view = binding.root
-        setContentView(view)
 
-        binding.choice1.setOnClickListener() {
-            startActivity(Intent(this, QuestChoice::class.java))
-            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
-            finish()
-        }
-        binding.choice2.setOnClickListener() {
-            startActivity(Intent(this, QuestChoice::class.java))
-            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
-            finish()
-        }
-        binding.choice3.setOnClickListener() {
-            startActivity(Intent(this, QuestChoice::class.java))
-            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
-            finish()
-        }
-    }
-}
-class EmotionChoice2Sad : AppCompatActivity() {
-    private lateinit var binding: EmotionChoice2HappyBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = EmotionChoice2HappyBinding.inflate(layoutInflater)
-        val view = binding.root
-        setContentView(view)
 
-        binding.choice1.setOnClickListener() {
-            startActivity(Intent(this, QuestChoice::class.java))
-            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
-            finish()
-        }
-        binding.choice2.setOnClickListener() {
-            startActivity(Intent(this, QuestChoice::class.java))
-            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
-            finish()
-        }
-        binding.choice3.setOnClickListener() {
-            startActivity(Intent(this, QuestChoice::class.java))
-            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
-            finish()
-        }
-    }
-}
-class QuestList : AppCompatActivity() {
-    private lateinit var binding: QuestListBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = QuestListBinding.inflate(layoutInflater)
-        val view = binding.root
-        setContentView(view)
 
-        binding.speechbubble.setOnClickListener() {
-            startActivity(Intent(this, NaviActivity::class.java))
-            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
-            finish()
-        }
-    }
-}
-class Header : AppCompatActivity() {
-    private lateinit var binding: HeaderBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = HeaderBinding.inflate(layoutInflater)
-        val view = binding.root
-        setContentView(view)
 
-        binding.wholerest.setOnClickListener() {
-            startActivity(Intent(this, NaviActivity::class.java))
-            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
-            finish()
-        }
-    }
-}
-class QuestChoice : AppCompatActivity() {
-    private lateinit var binding: QuestChoiceBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = QuestChoiceBinding.inflate(layoutInflater)
-        val view = binding.root
-        setContentView(view)
-
-        binding.choice1.setOnClickListener() {
-            startActivity(Intent(this, QuestList::class.java))
-            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
-            finish()
-        }
-        binding.choice2.setOnClickListener() {
-            startActivity(Intent(this, QuestList::class.java))
-            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
-            finish()
-        }
-        binding.choice3.setOnClickListener() {
-            startActivity(Intent(this, NaviActivity::class.java))
-            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
-            finish()
-        }
-    }
-}
-class QuestHistory : AppCompatActivity() {
-    private lateinit var binding: QuestHistoryBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = QuestHistoryBinding.inflate(layoutInflater)
-        val view = binding.root
-        setContentView(view)
-
-        binding.idDate.setOnClickListener() {
-            startActivity(Intent(this, QuestList::class.java))
-            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
-            finish()
-        }
-    }
-}
-class PostWrite : AppCompatActivity() {
-    private lateinit var binding: PostWriteBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = PostWriteBinding.inflate(layoutInflater)
-        val view = binding.root
-        setContentView(view)
-
-        binding.send.setOnClickListener() {
-            startActivity(Intent(this, QuestList::class.java))
-            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
-            finish()
-        }
-    }
-}
