+++ app/src/main/java/com/chobo/onrest/QuestList.kt
@@ -1,0 +1,22 @@
+package com.chobo.onrest
+
+import android.content.Intent
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.chobo.onrest.databinding.QuestListBinding
+
+class QuestList : AppCompatActivity() {
+    private lateinit var binding: QuestListBinding
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = QuestListBinding.inflate(layoutInflater)
+        val view = binding.root
+        setContentView(view)
+
+        binding.speechbubble.setOnClickListener() {
+            startActivity(Intent(this, NaviActivity::class.java))
+            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
+            finish()
+        }
+    }
+}
