+++ app/src/main/java/com/chobo/onrest/QuestionChoice.kt
@@ -1,0 +1,32 @@
+package com.chobo.onrest
+
+import android.content.Intent
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.chobo.onrest.databinding.QuestChoiceBinding
+
+class QuestChoice : AppCompatActivity() {
+    private lateinit var binding: QuestChoiceBinding
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = QuestChoiceBinding.inflate(layoutInflater)
+        val view = binding.root
+        setContentView(view)
+
+        binding.choice1.setOnClickListener() {
+            startActivity(Intent(this, QuestList::class.java))
+            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
+            finish()
+        }
+        binding.choice2.setOnClickListener() {
+            startActivity(Intent(this, QuestList::class.java))
+            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
+            finish()
+        }
+        binding.choice3.setOnClickListener() {
+            startActivity(Intent(this, NaviActivity::class.java))
+            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
+            finish()
+        }
+    }
+}
