+++ app/src/main/java/com/chobo/onrest/YourEmotion.kt
@@ -1,0 +1,27 @@
+package com.chobo.onrest
+
+import android.content.Intent
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.chobo.onrest.databinding.YourEmotionBinding
+
+class YourEmotion : AppCompatActivity() {
+    private lateinit var binding: YourEmotionBinding
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = YourEmotionBinding.inflate(layoutInflater)
+        val view = binding.root
+        setContentView(view)
+
+        binding.choice1.setOnClickListener() {
+            startActivity(Intent(this, QuestList::class.java))
+            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
+            finish()
+        }
+        binding.choice2.setOnClickListener() {
+            startActivity(Intent(this, EmotionChoice1::class.java))
+            overridePendingTransition( android.R.anim.fade_in, android.R.anim.fade_out )
+            finish()
+        }
+    }
+}
