+++ app/src/main/res/layout/emotion_choice2_sad.xml
